UniQScan/od_backend/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
flask-cors==4.0.1
flask-bcrypt==1.0.1
psycopg2-binary==2.9.9
pymupdf==1.24.5
python-dotenv==1.0.1
gunicorn==22.0.0
